remaining

+++ src/structs/data/low_level/LowLevelRebalanceData.sol
@@ -15,9 +15,9 @@ struct LowLevelRebalanceData {
     int256 protocolFutureRewardCollateral;
     int256 protocolFutureRewardBorrow;
     uint256 collateralPrice;
-    uint256 collateralTokenDecimals;
+    uint8 collateralTokenDecimals;
     uint256 borrowPrice;
-    uint256 borrowTokenDecimals;
+    uint8 borrowTokenDecimals;
     uint256 supplyAfterFee;
     uint256 totalAssets;
     uint16 targetLtvDividend;
